python/influx_pusher/requirements.txt
python-dotenv
influxdb_client
pandas